* 'master' of github.com:flatironinstitute/spikeforest2: (129 commits)
  writing docs
  ironclust fixed
  add timeseries widget
  changed file name ironclust_test.py to test0_ironclust.py
  spikesorters moved to spikesorters from spikeforest_analysis/sorters
  Update README.md
  Update README.md
  update test
  port option for sfbrowser and move old directories to old/
  add pylint requirement
  _keep_temp_files attribute on processor, and fix fortran order when writing raw data
  increment yass version
  update processing scripts
  yass .bin file storage method is debugged
  bug fixes
  added yass example data added yass direct test bypassing cairio
  fix _container bug in execute()
  update
  update readme
  update compute resource
  ...

+++ spikesorters/spyking_circus/spyking_circus.py
@@ -14,9 +14,12 @@ import shutil
 
 class SpykingCircus(mlpr.Processor):
     NAME = 'SpykingCircus'
-    VERSION = '0.2.0'
+    VERSION = '0.2.2'
     ENVIRONMENT_VARIABLES = [
         'NUM_WORKERS', 'MKL_NUM_THREADS', 'NUMEXPR_NUM_THREADS', 'OMP_NUM_THREADS']
+    ADDITIONAL_FILES = ['*.params']
+    CONTAINER = 'sha1://914becce45aec56a84dd1dd4bca4037b09c50373/02-12-2019/spyking_circus.simg'
+    CONTAINER_SHARE_ID = '69432e9201d0'  # place to look for container
 
     recording_dir = mlpr.Input('Directory of recording', directory=True)
     channels = mlpr.IntegerListParameter(
@@ -39,7 +42,7 @@ class SpykingCircus(mlpr.Processor):
     def run(self):
         code = ''.join(random.choice(string.ascii_uppercase)
                        for x in range(10))
-        tmpdir = os.environ.get('TEMPDIR', '/tmp')+'/ironclust-tmp-'+code
+        tmpdir = os.environ.get('TEMPDIR', '/tmp')+'/spyking-circus-tmp-'+code
 
         num_workers = os.environ.get('NUM_WORKERS', 1)
 
@@ -69,10 +72,12 @@ class SpykingCircus(mlpr.Processor):
             se.MdaSortingExtractor.writeSorting(
                 sorting=sorting, save_path=self.firings_out)
         except:
-            if os.path.exists(tmpdir):
-                shutil.rmtree(tmpdir)
+            if not getattr(self, '_keep_temp_files', False):
+                if os.path.exists(tmpdir):
+                    shutil.rmtree(tmpdir)
             raise
-        shutil.rmtree(tmpdir)
+        if not getattr(self, '_keep_temp_files', False):
+            shutil.rmtree(tmpdir)
 
 
 def spyking_circus(
@@ -125,7 +130,7 @@ def spyking_circus(
     elif file_name.endswith('.npy'):
         file_name = file_name[file_name.find('.npy')]
     np.save(join(output_folder, file_name),
-            recording.getTraces().astype('float32'))
+            recording.getTraces().astype('float32', order='F'))
 
     if detect_sign < 0:
         detect_sign = 'negative'
@@ -136,18 +141,23 @@ def spyking_circus(
 
     # set up spykingcircus config file
     with open(join(source_dir, 'config_default.params'), 'r') as f:
-        circus_config = f.readlines()
+        circus_config = f.read()
     if merge_spikes:
         auto = 1e-5
     else:
         auto = 0
-    circus_config = ''.join(circus_config).format(
+    circus_config = circus_config.format(
         float(recording.getSamplingFrequency()
               ), probe_file, template_width_ms, spike_thresh, detect_sign, filter,
         whitening_max_elts, clustering_max_elts, auto
     )
-    with open(join(output_folder, file_name + '.params'), 'w') as f:
-        f.writelines(circus_config)
+    params_file = join(output_folder, file_name + '.params')
+    with open(params_file, 'w') as f:
+        f.write(circus_config)
+
+    # with open(params_file) as ff:
+    #    print('CIRCUS CONFIG:')
+    #    print(ff.read())
 
     print('Running spyking circus...')
     t_start_proc = time.time()
